Fix MoveEntry parsing - Complete XGParser implementation
This commit completes the Go implementation of XGParser to match the Python reference implementation exactly.

Fixed Issues:
-------------

1. EngineStructBestMoveRecord.Dice field type (CRITICAL)
   - Changed: Dice [2]int8 → Dice [2]int32
   - Root cause: Python format '2l' means 2 int32s (8 bytes), not 2 int8s (2 bytes)
   - Impact: This 6-byte misalignment corrupted ALL MoveEntry fields:
     * AnalyzeL: Was 262143 → Now correctly 3
     * AnalyzeM: Was -1 → Now correctly 3
     * CompChoice: Was -1449787392 → Now correctly 1
     * CommentMove: Was 0 → Now correctly -1
     * All Eval arrays: Were zeros → Now show correct float values

2. Previously fixed (from earlier commits):
   - EngineStructDoubleAction.Crawford: int32 → int16 (2-byte misalignment)
   - HeaderMatchEntry.Invert: int8 → int32 (3-byte misalignment)
   - HeaderGameEntry padding: 1 byte → 3 bytes (2-byte misalignment)

Verification:
-------------
Complete diff between Python and Go output shows ONLY cosmetic differences:
- Boolean formatting (False vs false, True vs true)
- Float formatting (0.0 vs 0 for zero values)
- Float precision (Python shows float64, Go shows float32)
- UTF-8 byte string display differences

All numerical values are identical (within float32 precision).

Status:
-------
✅ All record types now parse correctly:
   - HeaderMatchEntry
   - HeaderGameEntry
   - CubeEntry (including EngineStructDoubleAction)
   - MoveEntry (including EngineStructBestMoveRecord)
   - FooterGameEntry
   - FooterMatchEntry

The Go parser is now a fully functional drop-in replacement for the Python implementation.

Files Modified:
--------------
- xgparser/xgstruct.go: Fixed Dice field type in EngineStructBestMoveRecord
- README.md: Updated status to COMPLETE, added comprehensive documentation
- FIXES_COMPLETE.md: Updated with all 4 fixes documented
- FINAL_SUMMARY.md: Added quick reference summary

Key Learning:
-------------
Python struct format notation is tricky - number prefixes apply ONLY to the immediately following character:
- '2l' = 2 int32s (8 bytes)
- 'l2l' = 1 int32 + 2 int32s = 3 int32s total (12 bytes)
- '9xxxx' = 9 bytes + 1 + 1 + 1 = 12 bytes total (NOT 13!)

Even a 1-byte misalignment causes cascading corruption of all subsequent fields.

+++ xgparser/xgstruct.go
@@ -200,7 +200,7 @@ func (e *EngineStructDoubleAction) FromStream(r io.Reader) error {
 // EngineStructBestMoveRecord represents best move analysis
 type EngineStructBestMoveRecord struct {
 	Pos       [26]int8
-	Dice      [2]int8
+	Dice      [2]int32 // Fixed: was [2]int8, Python uses 2l (2 int32s)
 	Level     int32
 	Score     [2]int32
 	Cube      int32
